Make Completion Work Even Better
Make it so that the Kubeconfig path can be completed, make client
creation lazy so we can pick up said kubeconfig, and add completion
functions to allow namespace overrides.
