+++ pkg/cmd/create/create_user.go
@@ -95,7 +95,10 @@ func createUser(factory *factory.Factory) *cobra.Command {
 			ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
 			defer cancel()
 
-			client := factory.Client()
+			client, err := factory.Client()
+			if err != nil {
+				return err
+			}
 
 			if err := o.validate(ctx, client); err != nil {
 				return err
