+++ pkg/cmd/flags/flags.go
@@ -19,10 +19,13 @@ package flags
 import (
 	"fmt"
 	"net/url"
+	"path/filepath"
 
 	"github.com/spf13/pflag"
 
 	"github.com/unikorn-cloud/kubectl-unikorn/pkg/cmd/errors"
+
+	"k8s.io/client-go/util/homedir"
 )
 
 type HostnameVar string
@@ -51,11 +54,13 @@ func (v HostnameVar) Type() string {
 }
 
 type UnikornFlags struct {
+	Kubeconfig        string
 	IdentityNamespace string
 	RegionNamespace   string
 }
 
 func (o *UnikornFlags) AddFlags(f *pflag.FlagSet) {
+	f.StringVar(&o.Kubeconfig, "kubeconfig", filepath.Join(homedir.HomeDir(), ".kube", "config"), "Kubernetes configuration file")
 	f.StringVar(&o.IdentityNamespace, "identity-namespace", "unikorn-identity", "Identity service namespace")
 	f.StringVar(&o.RegionNamespace, "region-namespace", "unikorn-region", "Region service namespace")
 }
