+++ pkg/cmd/get/get_user.go
@@ -196,7 +196,10 @@ func getUser(factory *factory.Factory) *cobra.Command {
 			ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
 			defer cancel()
 
-			client := factory.Client()
+			client, err := factory.Client()
+			if err != nil {
+				return err
+			}
 
 			if err := o.validate(ctx, client); err != nil {
 				return err
